improved efficiency for no breaking contigs

+++ src_reconcile/samba.sh
@@ -11,7 +11,7 @@ MIN_IDENTITY=0
 KMER=15
 SCORE=4
 NOBREAK="1"
-ALN_PARAM="map-ont -N 0 "
+ALN_PARAM="map-ont -N 1 "
 ALN_DATA="ont"
 POLISH_PARAM="--nano-raw"
 ALLOWED=""
@@ -168,12 +168,8 @@ QRYN=$ALN_DATA
 
 #minimap
 if [ ! -e scaffold_align.success ];then
-log "Aligning the reads to the contigs"
-  if [ $NOBREAK = "0" ];then
-    $MYPATH/../Flye/bin/flye-minimap2 -k $KMER -t $NUM_THREADS -x $ALN_PARAM $REF <(zcat -f $QRY | $MYPATH/fastqToFasta.pl ) 1> $REFN.$QRYN.paf.tmp 2>minimap.err && mv $REFN.$QRYN.paf.tmp $REFN.$QRYN.paf && touch scaffold_align.success && rm -f scaffold_split.success scaffold_filter.success || error_exit "minimap2 failed"
-  else
-    $MYPATH/../Flye/bin/flye-minimap2 -k $KMER -t $NUM_THREADS -x $ALN_PARAM <(ufasta one $REF |perl -ane 'BEGIN{$keep_on_ends='$KEEP_ON_ENDS';}{if($F[0] =~ /^>/){print}else{$sub=length($F[0])-2*$keep_on_ends;if($sub>10000){substr($F[0],$keep_on_ends,$sub,"N"x$sub); print $F[0],"\n";}else{print}}}')  <(zcat -f $QRY | $MYPATH/fastqToFasta.pl ) 1> $REFN.$QRYN.paf.tmp 2>minimap.err && mv $REFN.$QRYN.paf.tmp $REFN.$QRYN.paf && touch scaffold_align.success && rm -f scaffold_split.success scaffold_filter.success || error_exit "minimap2 failed"
-  fi
+  log "Aligning the reads to the contigs"
+  $MYPATH/../Flye/bin/flye-minimap2 -k $KMER -t $NUM_THREADS -x $ALN_PARAM $REF <(zcat -f $QRY | $MYPATH/fastqToFasta.pl ) 1> $REFN.$QRYN.paf.tmp 2>minimap.err && mv $REFN.$QRYN.paf.tmp $REFN.$QRYN.paf && touch scaffold_align.success && rm -f scaffold_split.success scaffold_filter.success || error_exit "minimap2 failed"
 fi
 
 if [ $NOBREAK = "0" ];then #we do not break in not instructed to or if the scaffolding sequence in another assembly
@@ -220,8 +216,11 @@ if [ $NOBREAK = "0" ];then #we do not break in not instructed to or if the scaff
   fi
 
 else
-  perl -ane '$F[5]=~s/:/s/g; print join("\t",@F),"\n";' $REFN.$QRYN.paf > $REFN.$QRYN.split.paf.tmp && mv $REFN.$QRYN.split.paf.tmp $REFN.$QRYN.split.paf && \
-  tr ':' 's' < $REF > $REFN.split.fa.tmp && mv $REFN.split.fa.tmp $REFN.split.fa
+  if [ ! -e scaffold_split.success ];then
+    perl -ane '$F[5]=~s/:/s/g; print join("\t",@F),"\n";' $REFN.$QRYN.paf > $REFN.$QRYN.split.paf.tmp && mv $REFN.$QRYN.split.paf.tmp $REFN.$QRYN.split.paf && \
+    tr ':' 's' < $REF > $REFN.split.fa.tmp && mv $REFN.split.fa.tmp $REFN.split.fa && touch scaffold_split.success && \
+    rm -f scaffold_split_align.success
+  fi
 fi
 
 if [ ! -e scaffold_filter.success ];then
